backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.17
pypdf==5.1.0
pikepdf==9.4.2
Pillow==11.0.0
reportlab==4.2.5
pdf2docx==0.5.8
pytesseract==0.3.13
python-pptx==1.0.2
openpyxl==3.1.5
pdfplumber==0.11.4
pdf2image==1.17.0
